% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_various.R
\name{cumhazinv}
\alias{cumhazinv}
\title{Inverse of the total cumulative hazard in simulation model}
\usage{
cumhazinv(u, X1, Z1, xi, rho, phi, pgen, lambda, v)
}
\description{
Internal function used used for simulating data
}
\keyword{internal}
