% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hessian.R
\name{hess}
\alias{hess}
\title{Hessian matrix as a function of (xi,rho,phi)}
\usage{
hess(par, nxi, nX, nZ, pset, datamat, psetTT, txi, tt, datInd)
}
\description{
Internal function used in estimation procedure
}
\keyword{internal}
