% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.r
\name{loglik}
\alias{loglik}
\title{Loglikelihood L as a function of (xi,rho,phi) to get starting values}
\usage{
loglik(par, nxi, nX, nZ, pset, datamat, psetTT, txi, tt, datInd)
}
\description{
Internal function used in estimation procedure
}
\keyword{internal}
