% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.r
\name{loglikopt}
\alias{loglikopt}
\title{Loglikelihood L as a function of (rho,phi) only, with xi known from previous iteration
Exactly the same as loglik except now xi is not an argument to the function so the first three
lines change}
\usage{
loglikopt(par, nxi, nX, nZ, pset, datamat, psetTT, txi, tt, datInd, opt, j)
}
\description{
Internal function used in estimation procedure
}
\keyword{internal}
