% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/likelihoods.r
\name{logliksteropt}
\alias{logliksteropt}
\title{Loglikelihood L* as a function of xi only, with rho and phi known from previous iteration
Only difference with L is denominator}
\usage{
logliksteropt(par, nxi, nX, nZ, pset, datamat, psetTT, txi, tt, datInd, opt, i)
}
\description{
Internal function used in estimation procedure
}
\keyword{internal}
