% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMCOD.R
\name{simMCOD}
\alias{simMCOD}
\title{Simulate survival data with multiple causes of death}
\usage{
simMCOD(n, xi, rho, phi, pgen, lambda, v, pUC)
}
\arguments{
\item{n}{Size of dataset to be generated}

\item{xi}{Log-ratio of the pure baseline hazards, assumed constant}

\item{rho}{Effect of the binary exposure X1(=Z1) on the pure hazard of the disease of interest}

\item{phi}{Effect of the binary exposure Z1(=X1) on the pure hazard of other diseases}

\item{pgen}{Vector of weights corresponding to each of the 6 states in the multistate model.}

\item{lambda}{Weibull scale parameter used to specify the pure baseline hazard of the disease of interest.}

\item{v}{Weibull shape parameter used to specify the pure baseline hazard of the disease of interest.}

\item{pUC}{Vector indicating the values in \code{pgen} corresponding to weight values assigned to the underlying cause of death.}
}
\value{
A data.frame as required by \code{\link[survMCOD]{survMCOD}} function.
Specifically, a dataset with one row per individual and the following variables:

\describe{
  \item{X1}{Binary exposure (to include in model for disease of interest).}
  \item{Z1}{Binary exposure equal to X1 (to include in model for other diseases).}
  \item{TimeEntry}{Time of entry into the study}
  \item{TimeExit}{Time of exit from the study}
  \item{Status}{Indicator of vital status, with Status=1 if individual died and Status=0 if individual is censored.}
  \item{Pi}{Weight indicating the proportion of the death attributed to the disease of interest (missing, i.e. "NA", if Status=0).}
  \item{UC}{The Underlying Cause Indicator, which is 1 if the individual died and the disease of interest was selected as
underlying cause of death and 0 otherwise (i.e. 0 if individual censored or if individual died but the disease of interest
was not selected as underlying cause of death).}
}
}
\description{
\code{simMCOD} simulates data from the multiple-cause model of Moreno-Betancur et al. (2017),
specifically under the same assumptions as in their simulation study.
}
\details{
The function can be used to generate data from any of the scenarios considered in the the main simulation settings
of Moreno-Betancur et al. (2017). See that reference for details.
}
\examples{

  datEx<-simMCOD(n=1000,xi=-1,rho=-2,phi=0,
          pgen=c(1,0,0.75,0.25,0.125,0.083),
          lambda=0.001,v=2,pUC=c(1,0.75))

  head(datEx)

}
\references{
Moreno-Betancur M, Sadaoui H, Piffaretti C, Rey G. Survival analysis with multiple causes of death:
Extending the competing risks model. Epidemiology 2017; 28(1): 12-19.
}
