% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcs_various.R
\name{stat}
\alias{stat}
\title{Non-parametric estimator of log-ratio of the baseline hazards}
\usage{
stat(datc, ind, nxi)
}
\description{
Internal function used to generate an starting value for this parameter for optimization
}
\keyword{internal}
