% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SurvM.R
\name{SurvM}
\alias{SurvM}
\title{Create a multiple-cause survival object}
\usage{
SurvM(time, time2, status, weight, type = c("right", "counting"))
}
\arguments{
\item{time}{For right censored data, this is the follow up time. For counting process data,
the first argument is the entry time.}

\item{time2}{For counting process data only, it is the exit time.}

\item{status}{The vital status indicator such that 0=alive and 1=dead.}

\item{weight}{The weight indicating the proportion of the death attributed to the disease of interest.
It should be NA if individual alive (status=0) and a number between 0 and 1 if dead (status=1).}

\item{type}{Character string specifying the type of censoring. Possible values are "right" and "counting".}
}
\value{
An object of class \code{SurvM}.
}
\description{
\code{SurvM} Creates a multiple-cause survival object, which is used as a response variable
in the model formula provided to \code{\link[survMCOD]{survMCOD}}.
}
\details{
This function needs to be used to in the formula argument of the \code{\link[survMCOD]{survMCOD}} function.

The user is referred to Moreno-Betancur et al. (2017), Piffaretti et al. (2016) and Rey et al. (2017)
for descriptions and discussions of various weight-attribution strategies to determine the \code{weight} argument.
}
\examples{

  datEx<-simMCOD(n=1000,xi=-1,rho=-2,phi=0,pgen=c(1,0,0.75,0.25,0.125,0.083),
          lambda=0.001,v=2,pUC=c(1,0.75))

  SurvM(time=datEx$TimeEntry, time2=datEx$TimeExit, status=datEx$Status,
  weight=datEx$Pi, type="counting")

}
\references{
Moreno-Betancur M, Sadaoui H, Piffaretti C, Rey G. Survival analysis with multiple causes of death:
Extending the competing risks model. Epidemiology 2017; 28(1): 12-19.

Piffaretti C, Moreno-Betancur M, Lamarche-Vadel A, Rey G. Quantifying cause-related mortality by
weighting multiple causes of death. Bulletin of the World Health Organization 2016; 94:870-879B.

Rey G, Piffaretti C, Rondet C, Lamarche-Vadel A, Moreno-Betancur M. Analyse de la mortalite par cause :
ponderation des causes multiples. Bulletin Epidemiologique Hebdomadaire, 2017; (1): 13-9.
}
